build: Make it easy to start the server standalone. Refs #1134
Currently, the server is started as a part of the `docker-compose`
command, which downloads docker images and runs a lot of other gradle
tasks, lengthening the development cycle.
Add a new script which will run the pre-requisite postgres db as a
docker image, expose the DB Port to the application and run the
`bootRun` task, which starts the server.

The `web-ui` can be run separately to use the `web-ui` for linting APIs