Scripts for migrating Database
The production environment for zally has been updated to run in a new
cluster. To support that, we need to migrate the existing data in the
api_review and rule_violation tables to a new Database.
Checking in the script application that will download the data to the
local machine for migrating it into a newer database